+++ src/main/java/de/samply/exporter/ExporterJob.java
@@ -112,6 +112,9 @@ public class ExporterJob {
                     projectBridgehead.setQueryState(finalQueryState);
                     projectBridgehead.setExporterResponse(exporterServiceResult.result());
                     projectBridgehead.setModifiedAt(Instant.now());
+                    if (finalQueryState == QueryState.FINISHED){
+                        projectBridgehead.setExporterDispatchCounter(projectBridgehead.getExporterDispatchCounter() + 1);
+                    }
                     projectBridgeheadRepository.save(projectBridgehead);
                     emailTemplateType.ifPresent(type -> sendEmail(projectBridgehead, type));
                     return Mono.empty();